Load Draconity's config from disc on engine initialization
Draconity's config is stored on disc in a predictable path. We need to know
which socket to connect to (and which secret to use, for authorization). For
now, don't allow custom Draconity configuration - just load this information
from the same file Draconity uses to configure itself.

In the long run, a system will need to be devised to handle Draconity requesting
a dynamic port. This will have to be handled in tandem with Draconity's
development.